requirements.txt
folium==0.19.4
haversine==2.9.0 
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==2.1.0
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
streamlit==1.41.1
streamlit_folium==0.24.0
